pyproject.toml

[project]
name = ""
version = "0.1.0"
description = ""
authors = [{ name = "Hju Kneyck Flores", email = "hjucode@gmail.com" }]
dependencies = ["websocket-client>=1.6.1", "rel>=0.4.9", "python-dotenv>=1.0.0"]
requires-python = ">=3.11"
license = { text = "MIT" }

[tool.pdm.dev-dependencies]
dev = ["pyright>=1.1.316", "black>=23.3.0", "ruff>=0.0.275", "isort>=5.12.0"]
